# connectors-leaf/api/openapi.yaml
openapi: 3.0.3
info:
  title: Connector & Evidence Service API
  description: Financial data connectors and evidence service for wealth management platforms
  version: 1.0.0
  contact:
    name: Platform Engineering Team
    email: platform@company.com
servers:
  - url: https://your-project-ref.supabase.co/functions/v1
    description: Production
  - url: http://localhost:54321/functions/v1
    description: Local Development
security:
  - BearerAuth: []
paths:
...
  /exceptions:
    get:
      summary: List exceptions
      description: Get list of data exceptions with filtering and pagination
      tags: [Exceptions]
      parameters:
        - name: entity_id
          in: query
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved]
            default: open
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of exceptions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Exception'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /exceptions/{id}/resolve:
    post:
      summary: Resolve exception
      description: Mark an exception as resolved with resolution notes
      tags: [Exceptions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolved_by, resolution_notes]
              properties:
                resolved_by:
                  type: string
                  format: uuid
                resolution_notes:
                  type: string
                resolution_action:
                  type: string
      responses:
        '200':
          description: Exception resolved successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Exception:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        exception_type:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, acknowledged, resolved]
        description:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
...